Create `commands` package and `core/` modules for NAP, Kueue and Pathways (#175)
* Create commands package and core/nap module

* Create Kueue and Pathways modules

* Move global variables to functions

* Add docstring

* Cleanup imports

* Add src/xpk/README.md
